# frontend/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LineSight",
    "description": "LineSight: AI-driven Semantic ETL platform for SMB apparel manufacturing. Parse messy Excel files, track SAM/DHU metrics, and ensure UFLPA compliance.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/datasources": {
      "post": {
        "tags": [
          "Data Sources"
        ],
        "summary": "Create Data Source",
        "description": "Create a new data source for a production line.",
        "operationId": "create_data_source_api_v1_datasources_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataSourceCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataSourceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Data Sources"
        ],
        "summary": "List Data Sources",
        "description": "List all data sources with mappings loaded.",
        "operationId": "list_data_sources_api_v1_datasources_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DataSourceResponse"
                  },
                  "title": "Response List Data Sources Api V1 Datasources Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasources/{data_source_id}": {
      "get": {
        "tags": [
          "Data Sources"
        ],
        "summary": "Get Data Source",
        "description": "Get data source by ID.",
        "operationId": "get_data_source_api_v1_datasources__data_source_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "data_source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Data Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataSourceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Data Sources"
        ],
        "summary": "Update Data Source",
        "description": "Update mappings or time column for a specific DataSource.",
        "operationId": "update_data_source_api_v1_datasources__data_source_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "data_source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Data Source Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataSourceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataSourceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Data Sources"
        ],
        "summary": "Delete Data Source",
        "description": "Delete a data source and all its associated schema mappings.\nThe database cascade should handle the mappings, but we'll be explicit if needed.",
        "operationId": "delete_data_source_api_v1_datasources__data_source_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "data_source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Data Source Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasources/line/{line_id}": {
      "get": {
        "tags": [
          "Data Sources"
        ],
        "summary": "Get Data Source By Line",
        "description": "Get data source for a specific production line.",
        "operationId": "get_data_source_by_line_api_v1_datasources_line__line_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "line_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Line Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/DataSourceResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Data Source By Line Api V1 Datasources Line  Line Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasources/by-line/{production_line_id}": {
      "get": {
        "tags": [
          "Data Sources"
        ],
        "summary": "Get Datasource By Line Explicit",
        "description": "Fetch DataSource config for a specific line.\nReturns 200 with null if not found (graceful init).\nAlias/Explicit endpoint for validate-mapping flow.",
        "operationId": "get_datasource_by_line_explicit_api_v1_datasources_by_line__production_line_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "production_line_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Production Line Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/DataSourceResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Datasource By Line Explicit Api V1 Datasources By Line  Production Line Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DataSourceCreate": {
        "properties": {
          "production_line_id": {
            "type": "string",
            "title": "Production Line Id"
          },
          "source_name": {
            "type": "string",
            "title": "Source Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "initial_mapping": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SchemaMappingCreate"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "production_line_id",
          "source_name"
        ],
        "title": "DataSourceCreate"
      },
      "DataSourceResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "production_line_id": {
            "type": "string",
            "title": "Production Line Id"
          },
          "source_name": {
            "type": "string",
            "title": "Source Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "time_column": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Column"
          },
          "time_format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Format"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "schema_mappings": {
            "items": {
              "$ref": "#/components/schemas/SchemaMappingResponse"
            },
            "type": "array",
            "title": "Schema Mappings",
            "default": []
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "production_line_id",
          "source_name",
          "description",
          "is_active",
          "created_at"
        ],
        "title": "DataSourceResponse"
      },
      "DataSourceUpdate": {
        "properties": {
          "time_column": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Column"
          },
          "time_format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Format"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "source_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Name"
          }
        },
        "type": "object",
        "title": "DataSourceUpdate",
        "description": "Schema for updating DataSource configuration."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SchemaMappingCreate": {
        "properties": {
          "column_map": {
            "additionalProperties": true,
            "type": "object",
            "title": "Column Map",
            "description": "Mapping from Excel columns to internal fields"
          },
          "extraction_rules": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Rules",
            "description": "Parsing rules (skip_rows, header_row, etc.)"
          },
          "reviewed_by_user": {
            "type": "boolean",
            "title": "Reviewed By User",
            "description": "Whether user has validated this mapping",
            "default": false
          },
          "user_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Notes"
          }
        },
        "type": "object",
        "required": [
          "column_map"
        ],
        "title": "SchemaMappingCreate"
      },
      "SchemaMappingResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "column_map": {
            "additionalProperties": true,
            "type": "object",
            "title": "Column Map"
          },
          "extraction_rules": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Rules"
          },
          "reviewed_by_user": {
            "type": "boolean",
            "title": "Reviewed By User"
          },
          "user_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "version",
          "is_active",
          "column_map",
          "extraction_rules",
          "reviewed_by_user",
          "user_notes",
          "created_at"
        ],
        "title": "SchemaMappingResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}